Move Layout and NotFound out of 'components' directory because they should not be included in generated ComponentFactory
